Add comments.

Convert mulhi_shr uses to use unsigned shift amount.

+++ src/runtime/hexagon_remote/host_malloc.cpp
@@ -176,8 +176,8 @@ void *halide_hexagon_host_malloc(size_t size) {
     ion_user_handle_t handle = 0;
     handle = ion_alloc(ion_fd, size, alignment, 1 << heap_id, ion_flags);
     if (handle < 0) {
-        __android_log_print(ANDROID_LOG_ERROR, "halide", "ion_alloc(%d, %d, %d, %d, %d) failed",
-                            ion_fd, size, alignment, 1 << heap_id, ion_flags);
+        __android_log_print(ANDROID_LOG_ERROR, "halide", "ion_alloc(%d, %lld, %lld, %d, %d) failed",
+                            ion_fd, (long long) size, (long long) alignment, 1 << heap_id, ion_flags);
         return NULL;
     }
     // Map the ion handle to a file buffer.
@@ -195,8 +195,8 @@ void *halide_hexagon_host_malloc(size_t size) {
     // Map the file buffer to a pointer.
     void *buf = mmap(NULL, size, PROT_READ | PROT_WRITE, MAP_SHARED, buf_fd, 0);
     if (buf == MAP_FAILED) {
-        __android_log_print(ANDROID_LOG_ERROR, "halide", "mmap(NULL, %d, PROT_READ | PROT_WRITE, MAP_SHARED, %d, 0) failed",
-                            size, buf_fd);
+        __android_log_print(ANDROID_LOG_ERROR, "halide", "mmap(NULL, %lld, PROT_READ | PROT_WRITE, MAP_SHARED, %d, 0) failed",
+                            (long long) size, buf_fd);
         close(buf_fd);
         if (!use_newer_ioctl) {
             ion_free(ion_fd, handle);
