+++ src/runtime/hexagon_remote/host_shim.cpp
@@ -2,7 +2,7 @@
 #include <stdio.h>
 #include <android/log.h>
 
-#include "bin/src/halide_hexagon_remote.h"
+#include "halide_hexagon_remote.h"
 
 typedef halide_hexagon_remote_handle_t handle_t;
 typedef halide_hexagon_remote_buffer buffer;
@@ -20,8 +20,8 @@ int halide_hexagon_remote_run(handle_t module_ptr, handle_t function,
     for (int i = 0; i < input_scalarsLen; i++) {
         int scalar_size = input_scalarsPtrs[i].dataLen;
         if (scalar_size > sizeof(scalar_t)) {
-            __android_log_print(ANDROID_LOG_ERROR, "halide", "Scalar argument %d is larger than %d bytes (%d bytes)",
-                                i, sizeof(scalar_t), scalar_size);
+            __android_log_print(ANDROID_LOG_ERROR, "halide", "Scalar argument %d is larger than %lld bytes (%d bytes)",
+                                i, (long long) sizeof(scalar_t), scalar_size);
             return -1;
         }
         memcpy(&scalars[i], input_scalarsPtrs[i].data, scalar_size);
